requirements.txt
Django~=4.0.6
psycopg2~=2.9.3
Faker~=13.15.1
djangorestframework~=3.13.1